crud abstraction

+++ common/database/referendum/crud.py
@@ -11,20 +11,20 @@ class CRUDBase(Generic[T]):
         self.model = model
 
     def create(self, db: Session, obj_in: schemas.BaseModel) -> T:
-        db_obj = self.model(**obj_in.dict())
+        db_obj = self.model(**obj_in.model_dump())
         db.add(db_obj)
         db.commit()
         db.refresh(db_obj)
         return db_obj
 
-    def get(self, db: Session, id: int) -> Optional[T]:
-        return db.query(self.model).filter(self.model.id == id).first()
+    def read(self, db: Session, obj_id: int) -> Optional[T]:
+        return db.query(self.model).filter(self.model.id == obj_id).first()
 
-    def get_multi(self, db: Session, skip: int = 0, limit: int = 100) -> List[T]:
+    def read_all(self, db: Session, skip: int = 0, limit: int = 100) -> List[T]:
         return db.query(self.model).offset(skip).limit(limit).all()
 
     def update(self, db: Session, db_obj: T, obj_in: schemas.BaseModel) -> T:
-        obj_data = obj_in.dict(exclude_unset=True)
+        obj_data = obj_in.model_dump(exclude_unset=True)
         for key, value in obj_data.items():
             setattr(db_obj, key, value)
         db.add(db_obj)
@@ -32,10 +32,11 @@ class CRUDBase(Generic[T]):
         db.refresh(db_obj)
         return db_obj
 
-    def delete(self, db: Session, id: int) -> None:
-        obj = db.query(self.model).get(id)
-        db.delete(obj)
-        db.commit()
+    def delete(self, db: Session, obj_id: int) -> None:
+        obj = db.get(self.model, obj_id)
+        if obj:
+            db.delete(obj)
+            db.commit()
 
 
 ### USERS ###
